代码优化

+++ src/main/java/org/FwTgt/botprofileEditor/dao/mapper/IBotAttributeMapper.java
@@ -8,14 +8,15 @@ import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
 public interface IBotAttributeMapper extends BaseMapper<BotAttribute> {
     @Insert("insert into botprofile_attributes values (#{botprofileId},#{attributeId})")
-    public void addRelation(@Param("botprofileId") int botprofileId,
-                            @Param("attributeId") int attributeId) throws Exception;
+    public void addRelation(@Param("botprofileId") BigInteger botprofileId,
+                            @Param("attributeId") BigInteger attributeId) throws Exception;
 
     @Select("select * from attributes where id in (" +
             "select attributeId from botprofile_attributes where botprofileId = #{botprofileId}" +
